removed incompatible google versions.
