+++ hosts/browserstack/android-devices/google.csv
@@ -14,5 +14,3 @@ Google Pixel 2	    ,8.0
 Google Pixel	    ,8.0
 Google Pixel	    ,7.1
 Google Nexus 6	    ,6.0
-Google Nexus 6	    ,5.0
-Google Nexus 5	    ,4.4
